+++ cmd/datawatch/main.go
@@ -86,7 +86,7 @@ import (
 )
 
 // Version is set at build time via -ldflags.
-var Version = "5.26.7"
+var Version = "5.26.8"
 
 var (
 	cfgPath    string